added support for ECMAScript5 properties.
this means on platforms like Node (and modern web browsers)
you can use dot (`.`) notation to get and set properties
that will pass through the KVO system.

this feature also dragged along `getPath` and `setPath` to
provide feature parity with the fact that `my.computed.property`
works with an arbitrary combination of properties and paths
using ECMAScript5 properties, but not using `get` or `set`.